Add reuse license compliance
This PR makes the project license declare licenses according to
https://reuse.software/ so that it passes the reuse lint.

Even though it is many affected files, the pr is scroll friendly and contains no logic.
It should hopefully be a quick general overview before merging it.

Basically, it adds license headers in spdx-format, as per the reuse-specification.
It also adds a missing reuse LICENSES file, ofl 1.1, and the .reuse/dep5 which takes care of the licening of json files etc.

Note: the reuse badge in the README will be green when the project is public

Signed-off-by: Josef Andersson <josef.andersson@digg.se>